Test the shared helper through RandomForest instead of TabDPT
The helper lives on AbstractModel and five wrappers call it, so the test does
not belong behind an optional GPU stack: `tabdpt` is an extra, and the fit
test beside it needs that package installed. RandomForest is a caller too and
needs nothing beyond sklearn, so the test runs wherever tabular tests run.

The `preserve_missing` flag test stays in test_tabdpt.py, since TabDPT is the
only caller that passes it.
