+++ tabular/tests/unittests/models/test_rf.py
@@ -1,5 +1,9 @@
 import copy
 
+import numpy as np
+import pandas as pd
+import pytest
+
 from autogluon.tabular.models.rf.rf_model import RFModel
 from autogluon.tabular.testing import FitHelper
 
@@ -63,3 +67,28 @@ def test_rf_binary_compile_onnx_no_config_bagging():
         FitHelper.fit_and_validate_dataset(
             dataset_name=dataset_name, fit_args=fit_args, compile=True, compiler_configs=compiler_configs
         )
+
+
+@pytest.mark.parametrize("preserve_missing", [False, True])
+def test_label_encode_categoricals_leaves_caller_frame_untouched(preserve_missing):
+    """The shared helper shares unencoded columns with the caller, so it must not write through them.
+
+    Encoded columns are replaced rather than edited, including the `preserve_missing`
+    write, which lands on the replacements.
+    """
+    values = np.array(["a", "b", "c"], dtype=object)[[0, 1, 2, 0, 1]].astype(object)
+    values[[1, 3]] = None
+    X = pd.DataFrame(
+        {
+            "num": np.arange(5, dtype=float),
+            "cat": pd.Series(values, dtype="category"),
+        }
+    )
+    X_before = X.copy(deep=True)
+
+    encoded = RFModel(problem_type="binary")._label_encode_categoricals(X, preserve_missing=preserve_missing)
+
+    assert X.equals(X_before)
+    assert list(X.dtypes) == list(X_before.dtypes)
+    # the untouched column is shared rather than copied, which is the point of the above
+    assert np.shares_memory(X["num"].to_numpy(), encoded["num"].to_numpy())
